fix(ir): loopSumElim trunc32 to signed int32 (negative coeffs stayed negative)
trunc32Stat wrapped intermediate Poly values via unsigned 32-bit, so a
negative coefficient like -23 became 4294967273 (2^32-23) when stored in
polyOf. The next arithmetic step read it back as a huge positive, tripped
the int32 overflow guard, and rejected the whole loop — so any reduction
with a negative coefficient was left running. p11_global_const_prop
(`s += foo(i)+bar(i) = s += 2i - 1`) ran 55026 steps instead of folding
to a constant.

Fix: wrap to the SIGNED int32 representation (negative stays negative),
matching the final-stage `trunc32` used on the closed-form sum.

Regression: p11 now 21 steps; all 10 quadratic probes PASS; perf 40/40,
stress 8/8, boundary 17/17; all perf assemble, no dup labels, no new
opcodes.

+++ src/ir/optim.cpp
@@ -925,7 +925,10 @@ long long safeMulStat(long long x, long long y, bool &bad) {
 long long trunc32Stat(long long x, bool &bad) {
   if (!bad && (x > INT32_MAX || x < INT32_MIN)) bad = true;
   if (bad) return 0;
-  return static_cast<long long>(static_cast<std::uint32_t>(x));
+  // wrap to the signed int32 representation (negative values stay negative,
+  // not their unsigned 2^32-1 form) so downstream arithmetic sees -23, not
+  // 4294967273.
+  return static_cast<long long>(static_cast<int>(static_cast<std::uint32_t>(x)));
 }
 }  // namespace
